+++ pkg/connectorrp/datamodel/connectormetadata.go
@@ -6,8 +6,8 @@
 package datamodel
 
 import (
-	"github.com/project-radius/radius/pkg/connectorrp/renderers"
 	"github.com/project-radius/radius/pkg/rp"
+	"github.com/project-radius/radius/pkg/ucp/resources"
 )
 
 // ConnectorMetadata represents internal DataModel properties common to all connector types.
@@ -19,7 +19,15 @@ type ConnectorMetadata struct {
 	// Stores action to retrieve secret values. For Azure, connectionstring is accessed through cosmos listConnectionString operation, if secrets are not provided as input
 	SecretValues map[string]rp.SecretValueReference `json:"secretValues,omitempty"`
 
-	RecipeData renderers.RecipeData
+	RecipeData RecipeData
+}
+
+type RecipeData struct {
+	Name               string
+	RecipeTemplatePath string
+	APIVersion         string
+	AzureResourceType  resources.KnownType
+	Resources          []string // Resource id's of the resources deployed by the recipe
 }
 
 // ConnectorRecipe is the recipe details used to automatically deploy underlying infrastructure for a connector
