+++ pkg/connectorrp/renderers/types.go
@@ -10,9 +10,9 @@ import (
 	"errors"
 
 	"github.com/project-radius/radius/pkg/armrpc/api/conv"
+	"github.com/project-radius/radius/pkg/connectorrp/datamodel"
 	"github.com/project-radius/radius/pkg/rp"
 	"github.com/project-radius/radius/pkg/rp/outputresource"
-	"github.com/project-radius/radius/pkg/ucp/resources"
 )
 
 const (
@@ -46,16 +46,16 @@ type RendererOutput struct {
 	Resources      []outputresource.OutputResource
 	ComputedValues map[string]ComputedValueReference
 	SecretValues   map[string]rp.SecretValueReference
-	RecipeData     RecipeData
+	RecipeData     datamodel.RecipeData
 }
 
-type RecipeData struct {
-	Name               string
-	RecipeTemplatePath string
-	APIVersion         string
-	AzureResourceType  resources.KnownType
-	Resources          []string // Resource id's of the resources deployed by the recipe
-}
+// type RecipeData struct {
+// 	Name               string
+// 	RecipeTemplatePath string
+// 	APIVersion         string
+// 	AzureResourceType  resources.KnownType
+// 	Resources          []string // Resource id's of the resources deployed by the recipe
+// }
 
 // ComputedValueReference represents a non-secret value that can accessed once the output resources
 // have been deployed.
